% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp.R
\name{abx_growth_gcomp}
\alias{abx_growth_gcomp}
\title{Function to do g-computation for antibiotics growth analysis}
\usage{
abx_growth_gcomp(
  data,
  laz_var_name = "mo3_haz",
  abx_var_name = "who_rec_abx",
  infection_var_name = "tac_shigella_attributable",
  severity_list = c("enroll_diar_blood", "enroll_diar_vom_days",
    "enroll_diar_fever_days", "enroll_diar_vom_num"),
  covariate_list = c("sex", "enr_age_months", "enr_haz", "final_quintile", "enroll_site"),
  site_var_name = "enroll_site",
  site_interaction = TRUE,
  age_var_name = "enr_age_months"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, perform traditional gcomputation. Else, perform longitudinal gcomputation}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{site_var_name}{name of site covariate in dataset (if applicable, else null)}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable, else NULL)}
}
\value{
\describe{
List of containing the following:
\item{\code{effect_inf_no_abx}}{numeric effect of infection on growth in subgroup that did \emph{not} receive antibiotics}
\item{\code{effect_inf_abx}}{numeric effect of infection on growth in subgroup that received antibiotics}
\item{\code{abx_0_inf_1}}{expected growth outcome in infected subgroup who did not receive abx}
\item{\code{abx_0_inf_0}}{expected growth outcome in uninfected subgroup who did not receive abx}
\item{\code{abx_1_inf_1}}{expected growth outcome in infected subgroup who received abx}
\item{\code{abx_1_inf_0}}{expected growth outcome in uninfected subgroup who receieved abx}
}
}
\description{
Function to do g-computation for antibiotics growth analysis
}
