% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcomp.R
\name{abx_growth_gcomp_case_control}
\alias{abx_growth_gcomp_case_control}
\title{Function to do g-computation for antibiotics growth analysis for case-control data}
\usage{
abx_growth_gcomp_case_control(
  data,
  laz_var_name = "mo3_haz",
  abx_var_name = "who_rec_abx",
  case_var_name = "case",
  severity_list = c("enroll_diar_blood", "enroll_diar_vom_days",
    "enroll_diar_fever_days", "enroll_diar_vom_num"),
  covariate_list = c("sex", "enr_age_months", "enr_haz", "final_quintile", "enroll_site"),
  covariate_list_control = NULL,
  site_var_name = "enroll_site",
  site_interaction = TRUE,
  age_var_name = "enr_age_months",
  outcome_type = "gaussian"
)
}
\arguments{
\item{data}{dataframe containing dataset to use for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{case_var_name}{name of binary exposure/case variable}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection).}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{covariate_list_control}{character vector containing names of baseline covariates for controls. If NULL, same covariate_list as cases}

\item{site_var_name}{name of site covariate in dataset (if applicable, else null)}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable, else NULL)}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}
}
\value{
\describe{
List of containing the following:
\item{\code{effect_inf_no_abx}}{effect of infection on growth in subgroup that did \emph{not} receive antibiotics vs heathy controls}
\item{\code{effect_inf_abx}}{effect of infection on growth in subgroup that received antibiotics vs healthy controls}
\item{\code{abx_0_inf_1}}{expected growth outcome in infected subgroup who did not receive abx}
\item{\code{abx_0_inf_0}}{expected growth outcome in healthy controls}
\item{\code{abx_1_inf_1}}{expected growth outcome in infected subgroup who received abx}
}
}
\description{
Function to do g-computation for antibiotics growth analysis for case-control data
}
