% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agaipw.R
\name{agaipw}
\alias{agaipw}
\title{Main function to get point estimates and confidence intervals using AIPW}
\usage{
agaipw(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name = NA,
  case_var_name = NA,
  site_var_name = NA,
  followup_var_names = NA,
  covariate_list,
  severity_list = NULL,
  pathogen_quantity_list = NULL,
  pathogen_attributable_list = NULL,
  no_etiology_var_name = NULL,
  outcome_type = "gaussian",
  sl.library.outcome = NULL,
  sl.library.outcome.2 = NULL,
  sl.library.outcome.case = NULL,
  sl.library.outcome.control = NULL,
  sl.library.treatment = c("SL.mean"),
  sl.library.infection = NULL,
  sl.library.missingness = NULL,
  sl.library.missingness.case = NULL,
  sl.library.missingness.control = NULL,
  case_control = FALSE,
  seed = 12345,
  v_folds = 5,
  return_models = FALSE,
  first_id_var_name = NULL,
  msm = FALSE,
  msm_var_name = NULL,
  msm_formula = NULL,
  ps_trunc_level = 0.01,
  parsimonious_propensity = TRUE
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable (for other diarrhea analyses; case_control = FALSE)}

\item{case_var_name}{name of variable indicating case (for case control analysis; case_control = TRUE)}

\item{site_var_name}{name of covariate for site (to exclude from propensity models)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, use AIPW without second stage regression}

\item{pathogen_quantity_list}{character vector containing name(s) of pathogen quantity variables in dataset}

\item{pathogen_attributable_list}{character vector containing name(s) of binary pathogen attributable variables in dataset. Used to identify diarrhea with no etiology if is.null(no_etiology_var_name)}

\item{no_etiology_var_name}{name of binary variable indicating diarrhea episodes with no known etiology. If null, use pathogen_attributable_list}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{sl.library.outcome}{list containing SuperLearner libraries to use for outcome regression}

\item{sl.library.outcome.2}{list containing SuperLearner libraries to use for 2nd stage outcome regression (baseline covariates only)}

\item{sl.library.outcome.case}{list containing SuperLearner libraries to use for case outcome regression}

\item{sl.library.outcome.control}{list containing SuperLearner libraries to use for control outcome regression}

\item{sl.library.treatment}{list containing SuperLearner libraries to use for antibiotic prescribing propensity model}

\item{sl.library.infection}{list containing SuperLearner libraries to use for infection propensity model}

\item{sl.library.missingness}{list containing SuperLearner libraries to use for outcome missingness model}

\item{sl.library.missingness.case}{list containing SuperLearner libraries to use for outcome missingness model (cases)}

\item{sl.library.missingness.control}{list containing SuperLearner libraries to use for outcome missingness model (controls)}

\item{case_control}{TRUE if case control analysis, FALSE for other diarrhea analysis. Default to FALSE}

\item{seed}{seed for reproducibility in SuperLearner cross-validation}

\item{v_folds}{number of cross-validation folds to use in SuperLearner}

\item{return_models}{boolean return SuperLearner models. Default FALSE.}

\item{first_id_var_name}{name of variable indicating unique participant identifier in data. Used to account for re-enrollment in study.}

\item{msm}{boolean indicating use of MSM for effect heterogeneity, default FALSE}

\item{msm_var_name}{name of variable to use for msm}

\item{msm_formula}{chatacter vector with formula to use for msm if msm TRUE}

\item{ps_trunc_level}{numeric value to truncate propensity scores \verb{< ps_trunc_level} or \verb{> 1 - ps_trunc_level}. Default to 0.01}

\item{parsimonious_propensity}{default TRUE for no etiology (TODO: FILL IN, IDK WHAT THIS MEANS. ALSO CAN WE RENAME IT? THIS IS LONG)}
}
\value{
List of class \code{agaipw_res} containing objects for AIPW result, AIPW models.
}
\description{
Note parameters include all options across other diarrhea and case control analyses
}
