% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw_case_control}
\alias{aipw_case_control}
\title{Function for AIPW estimates of effect of infection on growth- case-control analysis}
\usage{
aipw_case_control(
  data,
  laz_var_name,
  abx_var_name,
  case_var_name,
  site_var_name,
  covariate_list,
  severity_list,
  pathogen_quantity_list = NULL,
  outcome_type = "gaussian",
  sl.library.outcome.case = c("SL.glm"),
  sl.library.outcome.control = c("SL.glm"),
  sl.library.treatment = c("SL.mean"),
  sl.library.infection = c("SL.glm"),
  sl.library.missingness.case = c("SL.glm"),
  sl.library.missingness.control = c("SL.glm"),
  v_folds = 5,
  return_models = FALSE,
  first_id_var_name = NULL,
  msm = FALSE,
  msm_var_name = NULL,
  msm_formula = NULL,
  ps_trunc_level = 0.01
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{case_var_name}{name of variable indicating case (for case control analysis; case_control = TRUE)}

\item{site_var_name}{name of covariate for site (to exclude from propensity models)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, use AIPW without second stage regression}

\item{pathogen_quantity_list}{character vector containing name(s) of pathogen quantity variables in dataset}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{sl.library.outcome.case}{list containing SuperLearner libraries to use for case outcome regression}

\item{sl.library.outcome.control}{list containing SuperLearner libraries to use for control outcome regression}

\item{sl.library.treatment}{list containing SuperLearner libraries to use for antibiotic prescribing propensity model}

\item{sl.library.infection}{list containing SuperLearner libraries to use for infection propensity model}

\item{sl.library.missingness.case}{list containing SuperLearner libraries to use for outcome missingness model (cases)}

\item{sl.library.missingness.control}{list containing SuperLearner libraries to use for outcome missingness model (controls)}

\item{v_folds}{number of cross-validation folds to use in SuperLearner}

\item{return_models}{boolean return SuperLearner models. Default FALSE.}

\item{first_id_var_name}{name of variable indicating first_id in data. Used to account for re-enrollment in study.}

\item{msm}{boolean indicating use of MSM for effect heterogeneity, default FALSE}

\item{msm_var_name}{name of variable to use for msm}

\item{msm_formula}{chatacter vector with formula to use for msm if msm TRUE}

\item{ps_trunc_level}{numeric value to truncate propensity scores \verb{< ps_trunc_level} or \verb{> 1 - ps_trunc_level}. Default to 0.01}
}
\value{
List containing \code{aipw_case_control} object, models (if return_models = TRUE). \code{aipw_case_control} object contains dataframe with results, covariance matrix, standard errors.
}
\description{
Function for AIPW estimates of effect of infection on growth- case-control analysis
}
\keyword{internal}
