% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw_other_diarrhea}
\alias{aipw_other_diarrhea}
\title{Function for AIPW estimates of effect of infection on growth- single outcome regression}
\usage{
aipw_other_diarrhea(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  covariate_list,
  pathogen_quantity_list = NULL,
  pathogen_attributable_list = NULL,
  no_etiology_var_name = NULL,
  outcome_type = "gaussian",
  sl.library.outcome = c("SL.glm"),
  sl.library.treatment = c("SL.mean"),
  sl.library.infection = c("SL.glm"),
  sl.library.missingness = c("SL.glm"),
  v_folds = 5,
  return_models = FALSE,
  child_id_var_name = NULL
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable (for other diarrhea analyses; case_control = FALSE)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{pathogen_quantity_list}{character vector containing name(s) of pathogen quantity variables in dataset}

\item{pathogen_attributable_list}{character vector containing name(s) of binary pathogen attributable variables in dataset. Used to identify diarrhea with no etiology if is.null(no_etiology_var_name)}

\item{no_etiology_var_name}{name of binary variable indicating diarrhea episodes with no known etiology. If null, use pathogen_attributable_list}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{sl.library.outcome}{list containing SuperLearner libraries to use for outcome regression}

\item{sl.library.treatment}{list containing SuperLearner libraries to use for antibiotic prescribing propensity model}

\item{sl.library.infection}{list containing SuperLearner libraries to use for infection propensity model}

\item{sl.library.missingness}{list containing SuperLearner libraries to use for outcome missingness model}

\item{v_folds}{number of cross-validation folds to use in SuperLearner}

\item{return_models}{boolean return SuperLearner models. Default FALSE.}

\item{child_id_var_name}{name of variable indicating child_id in data. Used to account for re-enrollment in study.}
}
\value{
List containing \code{aipw_other_diarrhea} object, models (if return_models = TRUE). \code{aipw_other_diarrhea} object contains dataframe with results, covariance matrix, standard errors.
}
\description{
Function for AIPW estimates of effect of infection on growth- single outcome regression
}
\keyword{internal}
