% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw.R
\name{aipw_other_diarrhea_2}
\alias{aipw_other_diarrhea_2}
\title{Function for AIPW estimates of effect of infection on growth- two-stage outcome regression}
\usage{
aipw_other_diarrhea_2(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  site_var_name,
  covariate_list,
  severity_list,
  pathogen_quantity_list = NA,
  pathogen_attributable_list = NA,
  no_etiology_var_name = NA,
  outcome_type = "gaussian",
  sl.library.outcome = c("SL.glm"),
  sl.library.outcome.2 = c("SL.glm"),
  sl.library.treatment = c("SL.mean"),
  sl.library.infection = c("SL.glm"),
  sl.library.missingness = c("SL.glm"),
  v_folds = 5,
  return_models = FALSE,
  first_id_var_name = NULL,
  msm = FALSE,
  msm_var_name = NULL,
  msm_formula = NULL,
  ps_trunc_level = 0.01
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable (for other diarrhea analyses; case_control = FALSE)}

\item{site_var_name}{name of covariate for site (to exclude from propensity models)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, use AIPW without second stage regression}

\item{pathogen_quantity_list}{character vector containing name(s) of pathogen quantity variables in dataset}

\item{pathogen_attributable_list}{character vector containing name(s) of binary pathogen attributable variables in dataset. Used to identify diarrhea with no etiology if is.null(no_etiology_var_name)}

\item{no_etiology_var_name}{name of binary variable indicating diarrhea episodes with no known etiology. If null, use pathogen_attributable_list}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{sl.library.outcome}{list containing SuperLearner libraries to use for outcome regression}

\item{sl.library.outcome.2}{list containing SuperLearner libraries to use for 2nd stage outcome regression (baseline covariates only)}

\item{sl.library.treatment}{list containing SuperLearner libraries to use for antibiotic prescribing propensity model}

\item{sl.library.infection}{list containing SuperLearner libraries to use for infection propensity model}

\item{sl.library.missingness}{list containing SuperLearner libraries to use for outcome missingness model}

\item{v_folds}{number of cross-validation folds to use in SuperLearner}

\item{return_models}{boolean return SuperLearner models. Default FALSE.}

\item{first_id_var_name}{name of variable indicating first_id in data. Used to account for re-enrollment in study.}

\item{msm}{boolean indicating use of MSM for effect heterogeneity, default FALSE}

\item{msm_var_name}{name of variable to use for msm}

\item{msm_formula}{chatacter vector with formula to use for msm if msm TRUE}

\item{ps_trunc_level}{numeric value to truncate propensity scores \verb{< ps_trunc_level} or \verb{> 1 - ps_trunc_level}. Default to 0.01}

\item{sl.library.outcome.case}{list containing SuperLearner libraries to use for case outcome regression}

\item{sl.library.outcome.control}{list containing SuperLearner libraries to use for control outcome regression}
}
\value{
List containing \code{aipw_other_diarrhea_2} object, models (if return_models = TRUE). \code{aipw_other_diarrhea_2} object contains dataframe with results, covariance matrix, standard errors.
}
\description{
Function for AIPW estimates of effect of infection on growth- two-stage outcome regression
}
\keyword{internal}
