% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{average_agaipw_across_seeds}
\alias{average_agaipw_across_seeds}
\title{Helper function to average results across multiple seeds}
\usage{
average_agaipw_across_seeds(results_list)
}
\arguments{
\item{results_list}{list of results using same antibiotics levels and infection defintions, different seeds}
}
\value{
summary_df dataframe summarizing results across length(results_list) seeds
}
\description{
Helper function to average results across multiple seeds
}
