% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap_estimates}
\alias{bootstrap_estimates}
\title{Function to do n_boot bootstrap replicates and get bootstrap standard error}
\usage{
bootstrap_estimates(
  data,
  laz_var_name,
  abx_var_name,
  site_var_name,
  site_interaction,
  age_var_name,
  covariate_list,
  severity_list,
  n_boot,
  infection_var_name = NULL,
  case_var_name = NULL,
  case_control = FALSE,
  covariate_list_control = NULL,
  outcome_type = "gaussian",
  att = TRUE
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{site_var_name}{name of site covariate in dataset (if applicable, else null)}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable to add spline, else NULL)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, perform traditional gcomputation. Else, perform longitudinal gcomputation}

\item{n_boot}{number of bootstrap replicates to repeat}

\item{infection_var_name}{name of binary infection variable}

\item{case_var_name}{name of variable indicating case (only needed if case_control = TRUE)}

\item{case_control}{TRUE if case control analysis, FALSE otherwise}

\item{covariate_list_control}{covariate list for controls in case-control analysis, if not case control or covariate lists same then NULL}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{att}{boolean if effect should be estimated among people who would naturally get infection, default TRUE}
}
\value{
\describe{
List of containing bootstrap standard error and confidence intervals for each effect of interest:
\item{\code{se_abx}}{boostrap standard error for subgroup that received antibiotics}
\item{\code{lower_se_abx}}{lower bound of 95\% confidence interval for the subgroup that received antibiotics}
\item{\code{upper_se_abx}}{upper bound of 95\% confidence interval for the subgroup that received antibiotics}
\item{\code{se_no_abx}}{boostrap standard error for subgroup that did \emph{not} receive antibiotics}
\item{\code{lower_se_no_abx}}{lower bound of 95\% confidence interval for the subgroup that did \emph{not} receive antibiotics}
\item{\code{upper_se_no_abx}}{upper bound of 95\% confidence interval for the subgroup that that did \emph{not} receive antibiotics}
\item{\code{se_abx_0_inf_1}}{boostrap standard error for infected subgroup who did \emph{not} receive abx and was infected}
\item{\code{lower_ci_abx_0_inf_1}}{lower bound of 95\% confidence interval for the subgroup that did \emph{not} receive antibiotics and was infected}
\item{\code{upper_ci_abx_0_inf_1}}{upper bound of 95\% confidence interval for the subgroup that did \emph{not} receive antibiotics and was infected}
\item{\code{se_abx_0_inf_0}}{boostrap standard error for uninfected subgroup who did not receive abx}
\item{\code{lower_ci_abx_0_inf_0}}{lower bound of 95\% confidence interval for the subgroup that did \emph{not} receive antibiotics and was not infected}
\item{\code{upper_ci_abx_0_inf_0}}{upper bound of 95\% confidence interval for the subgroup that did \emph{not} receive antibiotics and was not infected}
\item{\code{se_abx_1_inf_1}}{boostrap standard error for infected subgroup who received abx}
\item{\code{lower_ci_abx_1_inf_1}}{lower bound of 95\% confidence interval for the subgroup that did receive antibiotics and was infected}
\item{\code{upper_ci_abx_1_inf_1}}{upper bound of 95\% confidence interval for the subgroup that did receive antibiotics and was infected}
\item{\code{se_abx_1_inf_0}}{boostrap standard error for uninfected subgroup who receieved abx - only in comparison to other diarrhea}
\item{\code{lower_ci_abx_1_inf_0}}{lower bound of 95\% confidence interval for the subgroup that did receive antibiotics and was not infected - only in comparison to other diarrhea}
\item{\code{upper_ci_abx_1_inf_0}}{upper bound of 95\% confidence interval for the subgroup that did receive antibiotics and was not infected - only in comparison to other diarrhea}
}
}
\description{
Function to do n_boot bootstrap replicates and get bootstrap standard error
}
\keyword{internal}
