% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_res}
\alias{combine_res}
\title{Helper function to combine multiple "aggcomp_res" objects into a CSV file of results}
\usage{
combine_res(res_list)
}
\arguments{
\item{res_list}{list of "aggcomp_res" objects}
}
\description{
Helper function to combine multiple "aggcomp_res" objects into a CSV file of results
}
