% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{impute_covariates}
\alias{impute_covariates}
\title{Helper function to impute select covariates with mean or mode}
\usage{
impute_covariates(
  data,
  imp_covariates,
  site_var_name = "site",
  imp_by_site = TRUE
)
}
\arguments{
\item{data}{dataset to impute}

\item{imp_covariates}{list of covariates to be imputed}

\item{site_var_name}{name of site variable if imputing by site}

\item{imp_by_site}{boolean indicating if imputing by site, default true}
}
\value{
dataframe with imputed data and flag columns indicating imputed values
}
\description{
Helper function to impute select covariates with mean or mode
}
