% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_utils.R, R/utils.R
\name{make_tables}
\alias{make_tables}
\title{Function to make and print descriptive tables to aid in interpreting gcomp}
\usage{
make_tables(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  site_var_name,
  site_interaction,
  age_var_name,
  severity_list,
  covariate_list
)

make_tables(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  site_var_name,
  site_interaction,
  age_var_name,
  severity_list,
  covariate_list
)
}
\arguments{
\item{data}{dataset used in analysis}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable}

\item{site_var_name}{name of site covariate in dataset (if applicable, else null)}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable, else NULL)}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, perform traditional gcomputation. Else, perform longitudinal gcomputation}

\item{covariate_list}{character vector containing names of baseline covariates}
}
\value{
list of descriptive tables

list of descriptive tables
}
\description{
Function to make and print descriptive tables to aid in interpreting gcomp

Function to make and print descriptive tables to aid in interpreting gcomp
}
