% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{one_boot}
\alias{one_boot}
\title{Function to do one bootstrap replicate}
\usage{
one_boot(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  site_var_name,
  site_interaction,
  age_var_name,
  covariate_list,
  severity_list,
  case_var_name = NULL,
  case_control = FALSE,
  covariate_list_control = NULL,
  outcome_type = "gaussian",
  att = TRUE
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable}

\item{site_var_name}{name of site covariate in dataset (if applicable, else null)}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable to add spline, else NULL)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection).}

\item{case_var_name}{name of variable indicating case (only needed if case_control = TRUE)}

\item{case_control}{TRUE if case control analysis, FALSE otherwise}

\item{covariate_list_control}{covariate list for controls in case-control analysis, if not case control or covariate lists same then NULL}

\item{outcome_type}{gaussian or binomial for continuous or binomial outcome}

\item{att}{boolean if effect should be estimated among people who would naturally get infection, default TRUE}
}
\value{
\describe{
List of containing the following for the bootstrapped dataset:
\item{\code{effect_inf_no_abx}}{numeric effect of infection on growth in subgroup that did \emph{not} receive antibiotics}
\item{\code{effect_inf_abx}}{numeric effect of infection on growth in subgroup that received antibiotics}
\item{\code{abx_0_inf_1}}{expected growth outcome in infected subgroup who did not receive abx}
\item{\code{abx_0_inf_0}}{expected growth outcome in uninfected subgroup who did not receive abx}
\item{\code{abx_1_inf_1}}{expected growth outcome in infected subgroup who received abx}
\item{\code{abx_1_inf_0}}{expected growth outcome in uninfected subgroup who receieved abx}
}
}
\description{
Function to do one bootstrap replicate
}
\keyword{internal}
