% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_cis}
\alias{plot_cis}
\title{Function to make confidence interval plots of results}
\usage{
plot_cis(results)
}
\arguments{
\item{results}{\code{aggcomp_res} object or list of \code{aggcomp_res} objects}
}
\value{
list of confidence interval ggplots
}
\description{
Function to make confidence interval plots of results
}
