% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.agaipw_res}
\alias{print.agaipw_res}
\title{Print the output of a \code{"agaipw_res"} object}
\usage{
\method{print}{agaipw_res}(x, ...)
}
\arguments{
\item{x}{An \code{"agaipw_res"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"agaipw_res"} object
}
