% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_utils.R, R/utils.R
\name{print.aggcomp_res}
\alias{print.aggcomp_res}
\title{Print the output of a \code{"aggcomp_res"} object}
\usage{
\method{print}{aggcomp_res}(x, ...)

\method{print}{aggcomp_res}(x, ...)
}
\arguments{
\item{x}{An \code{"aggcomp_res"} object.}

\item{...}{Other arguments (not used)}
}
\description{
Print the output of a \code{"aggcomp_res"} object

Print the output of a \code{"aggcomp_res"} object
}
