% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{aggcomp}
\alias{aggcomp}
\title{Main function to get point estimates and bootstrap confidence intervals for given data and parameters}
\usage{
aggcomp(
  data,
  laz_var_name,
  abx_var_name,
  infection_var_name,
  site_var_name,
  site_interaction,
  age_var_name,
  covariate_list,
  severity_list,
  n_boot = 1000,
  seed = 12345
)
}
\arguments{
\item{data}{dataframe containing dataset used for gcomp}

\item{laz_var_name}{name of growth outcome variable}

\item{abx_var_name}{name of binary antibiotic variable}

\item{infection_var_name}{name of binary infection variable}

\item{site_var_name}{name of site variable in dataset}

\item{site_interaction}{TRUE or FALSE indicating interaction between site and antibiotics}

\item{age_var_name}{name of age covariate in dataset (if applicable to add spline, else NULL)}

\item{covariate_list}{character vector containing names of baseline covariates}

\item{severity_list}{character vector containing names of severity-related covariates (post-infection). If NULL, perform traditional gcomputation. Else, perform longitudinal gcomputation}

\item{n_boot}{number of bootstrap replicates to repeat}

\item{seed}{seed to set for bootstrap}
}
\value{
List of class \code{aggcomp_res} containing point estimates, standard error, and 95\% confidence intervals for each effect of interest
}
\description{
Main function to get point estimates and bootstrap confidence intervals for given data and parameters
}
