% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/og_utils.R, R/utils.R
\name{plot_cis_subplots}
\alias{plot_cis_subplots}
\title{Function to make confidence interval plots of results with multiple subplots}
\usage{
plot_cis_subplots(
  results,
  abx_labels = NULL,
  inf_labels = NULL,
  outcome_labels = NULL
)

plot_cis_subplots(
  results,
  abx_labels = NULL,
  inf_labels = NULL,
  outcome_labels = NULL
)
}
\arguments{
\item{results}{\code{aggcomp_res} object or list of \code{aggcomp_res} objects}

\item{abx_labels}{label or list of labels to use for antibiotics variable (same length as results list)}

\item{inf_labels}{label or list of labels to use for infection variable (same length as results list)}

\item{outcome_labels}{label or list of labels to use for outcome variable (same length as results list)}
}
\value{
list of confidence interval ggplots

list of confidence interval ggplots
}
\description{
Function to make confidence interval plots of results with multiple subplots

Function to make confidence interval plots of results with multiple subplots
}
